\name{webmaps-package}
\Rdversion{1.1}
\alias{webmaps-package}
\alias{webmaps}
\docType{package}
\title{
Create web maps from spatial data, and import map tiles to R graphics.
}
\description{
  Leverage the power of javascript mapping library OpenLayers to make
  dynamic maps ready to deploy on a web site. Also, import map tiles
  into R graphics for pretty maps in R.
}
\details{
\tabular{ll}{
Package: \tab webmaps\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-08-06\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Barry Rowlingson

Maintainer: B.Rowlingson@lancaster.ac.uk
}
\keyword{ package }

